+++ api/api_athena/src/modules/minbar_predictor/ma_pred/grad_hunter/mghconf.h
@@ -53,8 +53,8 @@ public:
         return getKeyValue<int>(MA_HUNTER_ROOT + "TURN_POINT_OFFSET");
     }
 
-    real32 getSlopeZero() {
-        return getKeyValue<real32>(MA_HUNTER_ROOT + "SLOPE_ZERO");
+    real64 getSlopeZero() {
+        return getKeyValue<real64>(MA_HUNTER_ROOT + "SLOPE_ZERO");
     }
 
     real64 getFireOffSlope() {
